Handle each iframe individually when looking for the grid record count
By looking for the grid iframe using $("iframe") we were just collecting
up all the iframes and hunting through them at once. This should generally
be fine because the iframes are all using the same origin as this script,
however it seems that sometimes iframes are present with different origins
which results in cross-origin access errors in some browsers (chrome for
example) - I suspect a chrome extension was causing extra iframes to exist
on the page causing the errors to be raised. By treating each iframe
individually we can carefully catch any errors and ensure we only look for
the grid iframe we want.