Change how email addresses are processed and refactor
Prior to this change, if a user was logged in when using the UI
to make a download request they had no choice but to use the
email address associated with their user (the email input box was
not shown at all and the backend code looks for the email on the
user object).

Now the email input is always shown and is defaulted to the user's
email address if they are logged in. When the requests is received
by the backend it looks for the email input before it checks for a
logged in user.

Also, there's some refactoring in this commit with the removal of
some code and simplification of other bits (like using requests
instead of urllib).