Remove the anon functionality
Primarily this change has been made to make something work that was changed in a
previous commit - namely the email input box being shown even if the user is logged
in. The anon feature causes this email to always be used and anything entered by the
user is ignored (this is because the input change hooks aren't registered). To fix
this I've removed the anon functionality completely as we don't use it and I'm not
convinced it even properly works.