refactor: remove all functionality except database tables
BREAKING CHANGE: remove all functionality except database tables

ckanpackager is old and unstable and so is being deprecated. Most of this extension's functionality is based on a connection to a ckanpackager instance, so without that it's largely useless - however, some extensions like ckanext-statistics still depend on the download statistics from ckanpackager, so the database is being left intact so that these other extensions can have time to migrate the data to somewhere more sensible (or use this indefinitely).